feat(elements|ino-snackbar): rework snackbar (#431)
BREAKING CHANGE: remove types `primary` & `warning` and add new types `info` (default) & `success`; remove alignment properties

The snackbar has received a new design, which features the three cases info, success, and error. Also, as part of the new design, we have decided that the snackbar will now be displayed in the top right corner by default. Since we want to get away from offering properties that only make changes to the CSS in the background, we have removed the alignment properties and instead introduced CSS variables that can be used to configure the absolute position of the snackbar. For more information, have a look at its storybook page.