refactor(elements|ino-autocomplete): rework component to improve performance (#651)
BREAKING CHANGES: We've noticed performance issues when many options are provided. This is due to a limitation in the way slots are being handled within our framework. For this reason, the component has undergone a complete overhaul, resulting in a different API. Instead of custom option elements, the data must now be passed in as a string array or key-value pairs.