refactor(elements): remove `colorScheme` from some components (#756)
BREAKING CHANGES: As part of a broader design update, we have decided to remove the color-scheme from most of our components. The reason for this decision is that the color-scheme has grown historically and was intended to provide some configurability. We have discarded this configurability due to the increased focus on the elements as a design system and now want to ensure that everyone who uses the elements also uses them (semantically) correctly, resulting in a visually consistent application. We can only ensure this by limiting the configurability of the Elements. Another less important but not negligible reason is the greatly reduced maintainability, which allows us to work more effectively with our limited capacities. Since we have now disabled the shadow DOM for most components, you can still configure the elements using CSS. However, we cannot guarantee that future updates will not rename classes or restructure components.